+++ Tools/Sources/kd_decoder.h
@@ -36,7 +36,7 @@ public:
     void ResetDecoder();
     void InitDecoding(KdDecodable *dcodable);
 
-    long frame_num = 0; //number of decoded frame
+    uint frame_num = 0; //number of decoded frame
     int  uframe;       // reset on ResetDecoder(utterance)
 protected:
     // protected so classes which inherits also have access
